+++ tests/test_config.py
@@ -74,10 +74,11 @@ class ConfigTest(CoverageTest):
         assert not cov.config.branch
         assert cov.config.data_file == "delete.me"
 
-    def test_toml_config_file(self) -> None:
-        # A pyproject.toml file will be read into the configuration.
+    @pytest.mark.parametrize("filename", ["pyproject.toml", ".coveragerc.toml"])
+    def test_toml_config_file(self, filename: str) -> None:
+        # A pyproject.toml and .coveragerc.toml file will be read into the configuration.
         self.make_file(
-            "pyproject.toml",
+            filename,
             """\
             # This is just a bogus toml file for testing.
             [tool.somethingelse]
@@ -107,10 +108,11 @@ class ConfigTest(CoverageTest):
         assert cov.config.fail_under == 90.5
         assert cov.config.get_plugin_options("plugins.a_plugin") == {"hello": "world"}
 
-    def test_toml_ints_can_be_floats(self) -> None:
+    @pytest.mark.parametrize("filename", ["pyproject.toml", ".coveragerc.toml"])
+    def test_toml_ints_can_be_floats(self, filename: str) -> None:
         # Test that our class doesn't reject integers when loading floats
         self.make_file(
-            "pyproject.toml",
+            filename,
             """\
             # This is just a bogus toml file for testing.
             [tool.coverage.report]
@@ -262,6 +264,7 @@ class ConfigTest(CoverageTest):
         with pytest.raises(ConfigError, match=msg):
             coverage.Coverage()
 
+    @pytest.mark.parametrize("filename", ["pyproject.toml", ".coveragerc.toml"])
     @pytest.mark.parametrize(
         "bad_config, msg",
         [
@@ -298,9 +301,9 @@ class ConfigTest(CoverageTest):
             ('[tool.coverage.report]\nfail_under="s"', "couldn't convert to a float"),
         ],
     )
-    def test_toml_parse_errors(self, bad_config: str, msg: str) -> None:
+    def test_toml_parse_errors(self, filename: str, bad_config: str, msg: str) -> None:
         # Im-parsable values raise ConfigError, with details.
-        self.make_file("pyproject.toml", bad_config)
+        self.make_file(filename, bad_config)
         with pytest.raises(ConfigError, match=msg):
             coverage.Coverage()
 
@@ -329,10 +332,11 @@ class ConfigTest(CoverageTest):
         assert cov.config.branch is True
         assert cov.config.exclude_list == ["the_$one", "anotherZZZ", "xZZZy", "xy", "huh${X}what"]
 
-    def test_environment_vars_in_toml_config(self) -> None:
+    @pytest.mark.parametrize("filename", ["pyproject.toml", ".coveragerc.toml"])
+    def test_environment_vars_in_toml_config(self, filename: str) -> None:
         # Config files can have $envvars in them.
         self.make_file(
-            "pyproject.toml",
+            filename,
             """\
             [tool.coverage.run]
             data_file = "$DATA_FILE.fooey"
@@ -400,10 +404,11 @@ class ConfigTest(CoverageTest):
 
         self.assert_tilde_results()
 
-    def test_tilde_in_toml_config(self) -> None:
+    @pytest.mark.parametrize("filename", ["pyproject.toml", ".coveragerc.toml"])
+    def test_tilde_in_toml_config(self, filename: str) -> None:
         # Config entries that are file paths can be tilde-expanded.
         self.make_file(
-            "pyproject.toml",
+            filename,
             """\
             [tool.coverage.run]
             data_file = "~/data.file"
@@ -540,15 +545,16 @@ class ConfigTest(CoverageTest):
         with pytest.warns(CoverageWarning, match=msg):
             _ = coverage.Coverage()
 
-    def test_unknown_option_toml(self) -> None:
+    @pytest.mark.parametrize("filename", ["pyproject.toml", ".coveragerc.toml"])
+    def test_unknown_option_toml(self, filename: str) -> None:
         self.make_file(
-            "pyproject.toml",
+            filename,
             """\
             [tool.coverage.run]
             xyzzy = 17
             """,
         )
-        msg = r"Unrecognized option '\[tool.coverage.run\] xyzzy=' in config file pyproject.toml"
+        msg = f"Unrecognized option '\\[tool.coverage.run\\] xyzzy=' in config file {filename}"
         with pytest.warns(CoverageWarning, match=msg):
             _ = coverage.Coverage()
 
@@ -607,9 +613,10 @@ class ConfigTest(CoverageTest):
         with pytest.raises(ConfigError, match="No option 'foo' in section: 'xyzzy'"):
             config.get("xyzzy", "foo")
 
-    def test_exclude_also(self) -> None:
+    @pytest.mark.parametrize("filename", ["pyproject.toml", ".coveragerc.toml"])
+    def test_exclude_also(self, filename: str) -> None:
         self.make_file(
-            "pyproject.toml",
+            filename,
             """\
             [tool.coverage.report]
             exclude_also = ["foobar", "raise .*Error"]
@@ -958,10 +965,11 @@ class ConfigFileTest(UsingModulesMixin, CoverageTest):
                 coverage.Coverage(config_file="cov.toml")
 
     @pytest.mark.skipif(env.PYVERSION >= (3, 11), reason="Python 3.11 has toml in stdlib")
-    def test_no_toml_installed_pyproject_toml(self) -> None:
-        # Can't have coverage config in pyproject.toml without toml installed.
+    @pytest.mark.parametrize("filename", ["pyproject.toml", ".coveragerc.toml"])
+    def test_no_toml_installed_pyproject_toml(self, filename: str) -> None:
+        # Can't have coverage config in pyproject.toml/.coveragerc.toml without toml installed.
         self.make_file(
-            "pyproject.toml",
+            filename,
             """\
             # A toml file!
             [tool.coverage.run]
@@ -969,15 +977,16 @@ class ConfigFileTest(UsingModulesMixin, CoverageTest):
             """,
         )
         with mock.patch.object(coverage.tomlconfig, "has_tomllib", False):
-            msg = "Can't read 'pyproject.toml' without TOML support"
+            msg = f"Can't read '{filename}' without TOML support"
             with pytest.raises(ConfigError, match=msg):
                 coverage.Coverage()
 
     @pytest.mark.skipif(env.PYVERSION >= (3, 11), reason="Python 3.11 has toml in stdlib")
-    def test_no_toml_installed_pyproject_toml_shorter_syntax(self) -> None:
-        # Can't have coverage config in pyproject.toml without toml installed.
+    @pytest.mark.parametrize("filename", ["pyproject.toml", ".coveragerc.toml"])
+    def test_no_toml_installed_pyproject_toml_shorter_syntax(self, filename: str) -> None:
+        # Can't have coverage config in pyproject.toml/.coveragerc.toml without toml installed.
         self.make_file(
-            "pyproject.toml",
+            filename,
             """\
             # A toml file!
             [tool.coverage]
@@ -985,7 +994,7 @@ class ConfigFileTest(UsingModulesMixin, CoverageTest):
             """,
         )
         with mock.patch.object(coverage.tomlconfig, "has_tomllib", False):
-            msg = "Can't read 'pyproject.toml' without TOML support"
+            msg = f"Can't read '{filename}' without TOML support"
             with pytest.raises(ConfigError, match=msg):
                 coverage.Coverage()
 
@@ -1007,22 +1016,50 @@ class ConfigFileTest(UsingModulesMixin, CoverageTest):
             assert not cov.config.branch
             assert cov.config.data_file == ".coverage"
 
-    def test_exceptions_from_missing_toml_things(self) -> None:
+    @pytest.mark.parametrize("filename", ["pyproject.toml", ".coveragerc.toml"])
+    def test_exceptions_from_missing_toml_things(self, filename: str) -> None:
         self.make_file(
-            "pyproject.toml",
+            filename,
             """\
             [tool.coverage.run]
             branch = true
             """,
         )
         config = TomlConfigParser(False)
-        config.read("pyproject.toml")
+        config.read(filename)
         with pytest.raises(ConfigError, match="No section: 'xyzzy'"):
             config.options("xyzzy")
         with pytest.raises(ConfigError, match="No section: 'xyzzy'"):
             config.get("xyzzy", "foo")
         with pytest.raises(ConfigError, match="No option 'foo' in section: 'tool.coverage.run'"):
             config.get("run", "foo")
+
+    def test_coveragerc_toml_priority(self) -> None:
+        """Test that .coveragerc.toml has priority over pyproject.toml."""
+        self.make_file(
+            ".coveragerc.toml",
+            """\
+            [tool.coverage.run]
+            timid = true
+            data_file = ".toml-data.dat"
+            branch = true
+            """,
+        )
+
+        self.make_file(
+            "pyproject.toml",
+            """\
+            [tool.coverage.run]
+            timid = false
+            data_file = "pyproject-data.dat"
+            branch = false
+            """,
+        )
+        cov = coverage.Coverage()
+
+        assert cov.config.timid is True
+        assert cov.config.data_file == ".toml-data.dat"
+        assert cov.config.branch is True
 
 
 class SerializeConfigTest(CoverageTest):